move block cache from rpcdaemon to core (#1323)
small cleanups
