+++ silkworm/core/types/block_test.cpp
@@ -13,6 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
+
 #include "block.hpp"
 
 #include <catch2/catch.hpp>
